restructure into 3 files (Public, Protected, HTTPClient)
